requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
requests==2.32.3
python-dotenv==1.0.1